+++ Engine/Update/Public/Physics.h
@@ -4,17 +4,17 @@
 namespace physics {
     struct PhysicsInfo {
         bool physicsEnabled;
-        slurp::Vector2<int> position; // TODO: should make this a float vector
+        slurp::Vector2<float> position;
         slurp::Vector2<float> direction;
         float speed;
         // TODO: acceleration
 
         PhysicsInfo();
 
-        PhysicsInfo(const slurp::Vector2<int>& position);
+        PhysicsInfo(const slurp::Vector2<float>& position);
 
-        PhysicsInfo(const slurp::Vector2<int>& position, float speed);
+        PhysicsInfo(const slurp::Vector2<float>& position, float speed);
 
-        slurp::Vector2<int> getPositionUpdate(float dt);
+        slurp::Vector2<float> getPositionUpdate(float dt);
     };
 }
